+++ src-tauri/src/domain/repos/inventory_transaction_repo_trait.rs
@@ -2,7 +2,7 @@ use crate::common::error::AppError;
 use crate::domain::models::InventoryTransaction;
 
 pub trait InventoryTransactionRepoTrait: Send + Sync {
-    fn get_by_id(&self, id: i64) -> Result<Option<InventoryTransaction>, AppError>;
+    fn get_by_id(&self, id: i32) -> Result<Option<InventoryTransaction>, AppError>;
     fn create(&self, tx: &InventoryTransaction) -> Result<(), AppError>;
     fn list_for_product(&self, upc: String) -> Result<Vec<InventoryTransaction>, AppError>;
     fn list_for_operator(&self, operator_mdoc: i32) -> Result<Vec<InventoryTransaction>, AppError>;
@@ -11,12 +11,12 @@ pub trait InventoryTransactionRepoTrait: Send + Sync {
     fn list(&self) -> Result<Vec<InventoryTransaction>, AppError>;
     fn search(
         &self,
-        limit: i64,
-        offset: i64,
+        limit: i32,
+        offset: i32,
         date: Option<String>,
         search: Option<String>,
     ) -> Result<Vec<(InventoryTransaction, String, String)>, AppError>;
-    fn count(&self, date: Option<String>, search: Option<String>) -> Result<i64, AppError>;
+    fn count(&self, date: Option<String>, search: Option<String>) -> Result<i32, AppError>;
     fn create_with_tx(
         &self,
         a: &InventoryTransaction,
