+++ src-tauri/src/interface/dto/customer_transaction_dto.rs
@@ -21,14 +21,14 @@ pub struct CustomerTransactionDto {
 #[derive(Serialize)]
 pub struct CustomerTransactionSearchResult {
     pub items: Vec<CustomerTransactionSearchRow>,
-    pub total_count: u32,
+    pub total_count: i32,
 }
 
 #[derive(Serialize)]
 pub struct CustomerTransactionSearchRow {
     pub transaction: CustomerTransactionDto,
     pub operator_name: String,
-    pub spent: i64,
+    pub spent: i32,
 }
 
 #[cfg(test)]
